export PYTHONPATH for configure
bitbake does not export variables with "prepend". For example:
export PYTHONPATH:prepend = "/my/path:" does not actually export
PYTHONPATH to the environment.

Therefore, when the line for exporting PYTHONPATH was changed to use the
prepend tag, it was not longer being exported and
`ament-cmake-core-native` could no longer build in yocto.

(This feels like a bug in bitbake to me, but there is a way to work
around it.)

Related-commit: 9697407b62b339a3755b8b1399261e652aef1e94